Задача №7: Условие вывода данныхЗадание Необходимо написать программу, где бы пользователю предлагалось ввести с клавиатуры число в переменную number. В программе должны присутствовать константы X, Y, Z. Программа должна сравнивать введенное значение с клавиатуры со значением констант и вывести на экран сообщение "Данное значение имеется в константах" если number=X или Y или Z. Если переменная отличается от констант, нужно вывести оповещение "Такой константы нет!"

+++ src/Task_6.java
@@ -3,14 +3,14 @@ public class Task_6 {
     public static void main (String[] args) {
         Scanner in = new Scanner(System.in);
         System.out.print("Введите x: ");
-        int x = in.nextInt();
+        double x = in.nextInt();
         System.out.print("Введите y: ");
-        int y = in.nextInt();
+        double y = in.nextInt();
         System.out.print("Введите z: ");
-        int z = in.nextInt();
-        int avr = (x+y+z)/3;
+        double z = in.nextInt();
+        double avr = (x+y+z)/3;
         System.out.println("Среднее арифметическое чисел x, y, z: " + avr);
-        int i = avr/2;
+        double i = avr/2;
         if (i > 3)
         {
             System.out.println("Программа выполнена корректно");
